local-requirements.txt
hemlock-demographics==0.0.4
hemlock-survey>=0.0.33
ipykernel>=5.3.4
pyyaml>=5.3.1
xgboost>=1.2.1
gshap>=0.0.3
scikit-learn==0.23.2
statsmodels==0.12.1
tpot==0.11.6